marketing_anal/etl/requirements.txt
google-cloud-firestore==2.13.1
pandas==2.1.3
python-dotenv==1.0.0
firebase-admin==6.3.0
